Move tab logic into vscode/showTextDocument
